feat: Dump in-flight requests in case of a crash
The inproc and breakpad backends will invoke the default transports dump method,
that method will first check that it is actually the transport responsible,
and will then dump all the envelopes in matching tasks.
The bg worker is changed in a way that it only pops tasks *after* they are completed.
That way, in case of a race, we end up with duplicate events rather than dropped events.