Add sentry_reset_backend API
We have a third-party dynamic library (.NET Core runtime) that is loaded far into our application's runtime which, on load, steals all signal handlers. On Linux this works out fine, they are responsible and call whatever signal handler was there before them. No such luck on Windows. Forking and patching the .NET Core runtime is quite a bit more work, and we may run into this kind of issue in the future, so our internal fix is to add this API, and hopefully we can get this upstreamed and avoid forking sentry-native altogether.