fix: Clone mutable scope data when applying to event
The objects on the scope are mutable, which means they can re-allocate at any time.
This could potentially lead to segfaults when concurrently serializing an event object
that has a reference to such a scope object.
Now we shallowly clone those objects into the event which should guarantee they do not
move.