AIX support for module finder
Uses the loader subsystem, which is slightly awkward.

While we could use procfs, this is unavailable under PASE and thus
we'll have to rely on something else instead.

This uses code I originally wrote for Mono, adapted for the purpose.
Mono is under the same license, so this is compatible legally and I
also wrote it.