feat(tracing): Move all performance monitoring code behind a compile flag (#645)
This moves both the external and the internal parts of the 
API related to performance monitoring (/tracing) behind a 
compile flag. This is to make it easier for us to iterate on 
the API, and to potentially break it or leave it in a 
compilable but incomplete state. 

The flag is also enabled by default when building the 
example to speed up local testing during development.