fix: Read Windows Version from Registry (#623)
We were previously relying on the `kernel32.dll` version, as does
Crashpad. However that version can lag behind OS releases, which can
manifest itself as failing integration tests, as the version does not
match the one returned from python `platform.version()`.
This tries to use `ntoskrnl.exe` as the kernel version, but falls back to
`kernel32.dll` as before in case that does not work.