ref/ci: Update black as well as Windows jobs to unbreak CI (#691, #692)
This aggressively bumps black up to a new version to fix
our linter jobs in CI. This is fine because the use of Python 
is exclusively confined to integration tests, which makes
the impact of this update fairly small.

This change also updates the CI to remove the job running
tests on a Windows Server 2016 environment as it is now
deprecated and can no longer be used. A job running
the same tests on Windows Server 2019 has been created
to replace the removed job, as `windows-latest` now
uses Windows Server 2022.