fix(ci): change CC and CXX for macOS clang config
The new virtual-environment image
https://github.com/actions/virtual-environments/releases/tag/macOS-11%2F20220627.1
defaults clang and clang++ in the PATH to 13.0.0 which currently fails
in the CI because there is no "detect_leaks" support.

Clang/LLVM 13.0.1 is also packaged in that image but only accessible via
'$(brew --prefix llvm@13)/bin/clang'.

It is unclear whether this fixes the issue, but in the previous image
we used 13.0.1 successfully with ASAN.