fix: open the file lock on "UNIX" with O_RDRW (#791)
For some reason, the file lock on UNIX was always opened using O_RDONLY.
This doesn't make sense in combination with O_TRUNC, since this is a
write operation. On some systems (customer-scenario was AWS EFS) this
will lead to an EBADF if we try to flock that descriptor. The docs
consider this combination undefined.

This change also gets rid of the AIX compile conditional, since this
should work on all systems now.

I also added a warning to the run-initialization in case we get an error
when trying to lock.

macOS-11 CI runner needed to be updated to llvm15:

https://github.com/actions/runner-images/blob/macOS-11/20230117.2/images/macos/macos-11-Readme.md