fix: use void for all empty parameter-lists in C functions
This is one of many aspects where C and C++ standards deviate:

An empty parameter list in C++ is equivalent to a void parameter list. But in C an empty parameter list meant: you can give any number of any types as parameters (difference between declaration and definition ignored for brevity).

With C11 this "feature" became obsolete: empty parameter list are no longer standard-conforming C. Since some of those are in our API-surface we should fix this.