fix: specify correct dependencies for CMake client projects... (#926)
...which use a system-provided breakpad (instead of our vendored fork).

The problem [here](https://github.com/getsentry/sentry-native/issues/877)
was that people who introduce sentry-native via CMake `find_package()`
will get insufficient dependencies, which leads to configuration errors
in the client CMake project.

There are two aspects to this problem:
* if the user builds sentry as a shared library, it shouldn't be
necessary to specify the dependencies. This can be fixed by defining
`breakpad` as a `PRIVATE` dependency. This should also fix the Gentoo
issue because it uses sentry as a shared library afaict.
* if the user builds sentry as a static library, then we must stay with
the `PUBLIC` dependency, but we also need to correctly search for
`breakpad`, `libcurl`, and `pthread` in the context of the client project.