fix the UAF that the basic unit test revealed in the proposed implementation.
The fix also visualizes the event_id generation and capture validity discrepancy much better.

The bug in the proposed implementation was retrieving the event_id from the envelope after it had been captured. In this case we have no control over the validity of the envelope because the capture function takes ownership and frees the envelope. Retrieving the event_id after the capture is thus a race that leads to a UAF.

We can only safely retrieve the event_id from the moment where we create the envelope from the event (which is also where the event_id is generated). Of course after this a few things can still go wrong, which is why we cannot maintain the event_id as return value, but only return it if we reached the capture.

In all other cases we return a nil UUID, as specified in the comment.