API Change / Addition to sentry_capture_minidump (#1138)
* change the return type of sentry_capture_minidump to have information about the event

* add InsertNewlineAtEOF clang-format config...

...for users whose editor does not pick up `.editorconfig`.

* cleanup CHANGELOG.md

* extract `Unreleased` after merging master (containing `0.7.20`)

* add unit tests to cover basic error cases

* fix the UAF that the basic unit test revealed in the proposed implementation.

The fix also visualizes the event_id generation and capture-validity discrepancy much better.

The bug in the proposed implementation was retrieving the event_id from the envelope after it had been captured. In this case, we have no control over the validity of the envelope because the capture function takes ownership and frees the envelope. Retrieving the event_id after the capture is thus a race that leads to a UAF.

We can only safely retrieve the event_id from the moment we create the envelope from the event (also where the event_id is generated). Of course, a few things can still go wrong after this, so we cannot maintain the event_id as a return value; we can only return it if we reach the capture.

In all other cases, we return a nil UUID, as specified in the comment.

* Ensure we free the envelope when we fail to load the minidump attachment.

* Ensure we free the event in case `sentry_init()` hasn't been called.

* Since we inverted the function nesting sequence, keep track of `event` inside the `options` scope. Ensure to track even for failed envelope-item and failed envelope creation.

* fix Windows tests (which failed because we used the `wchar_t` string from the concatenated path as a parameter to `sentry_capture_minidump()`, which only takes UTF-8 `char` strings).

* ensure early exit from options scope decrements the refcount.

Also: freeing envelopes decreases the reference count of the event from which they were created.

* skip capture_minidump_basic on the  Android emulator since it can't work there.

* add a changelog for the fix to the pre-PR implementation

---------

Co-authored-by: Zach Danz <zach.danz@zenimaxonline.com>
Co-authored-by: Mischan Toosarani-Hausberger <mischan@abovevacant.com>