fix lock-order inversion in `sentry_start_session()`
This will typically not hit because the inversion is relevant with the lock acquisition in `sentry_init()` and these two are typically not called from two different threads at the same time.

However, consistent lock-order is such a basic sanity step and will guard us from future functions that also require scope- and options-locks.