re-enable all crashing tests in the http/stdout integration tests
I falsely assumed an issue in the signal handler itself because the internal check triggered by crashpad_handler forking coincided with DEADLYSIGNAL.

Turns out DEADLYSIGNAL is handled in TSAN like in ASAN, i.e. non-fatal (puh).

The culprit for crashpad is the fork during initialization, which breaks with TSAN invariants.